编写课件编辑页-4

+++ models/Children.php
@@ -9,7 +9,7 @@
 namespace app\models;
 
 use yii\base\Behavior;
-use yii\helpers\ArrayHelper;
+use Yii;
 use yii\db\ActiveRecord;
 
 /**
@@ -19,6 +19,8 @@ use yii\db\ActiveRecord;
  * @property string $childClass
  * @property \yii\db\ActiveRecord[] $children
  * @property \yii\db\ActiveRecord[] $_children
+ * @property \yii\db\ActiveRecord[] $_oldChildren
+ * @property \yii\db\ActiveRecord $owner
  *
  */
 class Children extends Behavior
@@ -33,15 +35,23 @@ class Children extends Behavior
             ActiveRecord::EVENT_BEFORE_VALIDATE => 'validate',
             ActiveRecord::EVENT_BEFORE_INSERT => 'save',
             ActiveRecord::EVENT_BEFORE_UPDATE => 'save',
-            // TODO: before delete
+            ActiveRecord::EVENT_BEFORE_DELETE => 'delete',
         ];
     }
 
+    /**
+     * @return ActiveRecord[]
+     */
     public function getChildren()
     {
         if ($this->_children === null) {
             $class = $this->childClass;
-            $this->_children = $class::findAll(Tree::children($this->owner->primaryKey));
+
+            $this->_children = [];
+            foreach (Tree::children($this->owner->primaryKey) as $id) {
+                $this->_children[]= $class::findOne($id);
+            }
+            $this->_oldChildren = $this->_children;
         }
         return $this->_children;
     }
@@ -51,33 +61,39 @@ class Children extends Behavior
      */
     public function setChildren($newChildren)
     {
-        $this->_oldChildren = $this->_children;
         $this->_children = $newChildren;
     }
 
-    public function validate()
+    public function validate($event)
     {
         if (!$this->_children) {
             return true;
         }
         foreach ($this->_children as $child) {
             if (!$child->validate()) {
+                $event->isValid = false;
                 return false;
             }
         }
         return true;
     }
 
-    public function save()
+    public function save($event)
     {
+        if ($this->_children === null) {
+            return true;
+        }
+
         $newList = [];
+
+        // save children data
         foreach ($this->_children as $child) {
             $child->save();
             $newList[] = $child->primaryKey;
         }
 
         // remove not used old children
-        if ($this->_oldChildren) {
+        if ($this->_oldChildren && $event->name == ActiveRecord::EVENT_BEFORE_UPDATE) {
             foreach ($this->_oldChildren as $child) {
                 if (!in_array($child->primaryKey, $newList)) {
                     $child->delete();
@@ -87,6 +103,14 @@ class Children extends Behavior
 
         // save new children list
         Tree::setChildren($this->owner->primaryKey, $newList);
+
         return true;
+    }
+
+    public function delete($event)
+    {
+        foreach ($this->_children as $child) {
+            $child->delete();
+        }
     }
 }
